+++ docs/CHANGELOG.md
@@ -1,8 +1,8 @@
 # 1.30.5
 
-- telling the user that their account will now be disabled.
+- Telling the user that their account will now be disabled.
 
-- analytics for adding documents...
+- Analytics for adding documents...
 
 # 1.30.4